changed to unsigned long long int for parameter operator""

+++ klib/units.hpp
@@ -162,8 +162,8 @@ namespace klib::time {
      * @brief Operators for all the time units
      * 
      */
-    constexpr ns operator"" _ns(const uint64_t value) {
-        // get rid of the uint64_t straight away
+    constexpr ns operator"" _ns(const unsigned long long int value) {
+        // get rid of the unsigned long long int straight away
         return ns(static_cast<uint32_t>(value));
     }
 
@@ -179,8 +179,8 @@ namespace klib::time {
         return (value / conversion_factor<ns, s>());
     }
 
-    constexpr us operator"" _us(const uint64_t value) {
-        // get rid of the uint64_t straight away
+    constexpr us operator"" _us(const unsigned long long int value) {
+        // get rid of the unsigned long long int straight away
         return us(static_cast<uint32_t>(value));
     }
 
@@ -196,8 +196,8 @@ namespace klib::time {
         return (value / conversion_factor<us, s>());
     }
 
-    constexpr ms operator"" _ms(const uint64_t value) {
-        // get rid of the uint64_t straight away
+    constexpr ms operator"" _ms(const unsigned long long int value) {
+        // get rid of the unsigned long long int straight away
         return ms(static_cast<uint32_t>(value));
     }
 
@@ -213,8 +213,8 @@ namespace klib::time {
         return (value / conversion_factor<ms, s>());
     }
 
-    constexpr s operator"" _s(const uint64_t value) {
-        // get rid of the uint64_t straight away
+    constexpr s operator"" _s(const unsigned long long int value) {
+        // get rid of the unsigned long long int straight away
         return s(static_cast<uint32_t>(value));
     }
